Goblins for state transition systems.
Goblins are structured mutators for a type; structured in that the nature of
that mutation is goverened by their genome. GHC Generics provide a way to map
this genome to a Goblin.

The intention is to try to 'breed' goblins for a state transition system, such
that a goblin is bred to cause specific types of failure in that system.